+++ packages/expo-video/android/src/main/java/expo/modules/video/listeners/VideoManagerListener.kt
@@ -1,0 +1,11 @@
+package expo.modules.video.listeners
+
+import expo.modules.video.VideoView
+
+// TODO: Add more as needed
+interface VideoManagerListener {
+  fun onVideoViewRegistered(videoView: VideoView, allVideoViews: Collection<VideoView>) {}
+  fun onVideoViewUnregistered(videoView: VideoView, allVideoViews: Collection<VideoView>) {}
+  fun onAppBackgrounded() {}
+  fun onAppForegrounded() {}
+}